Adds STANDARD_CI_REPOSITORY_TYPE
This is the name of the SCM the repo is using.

Although there might be better names than `STANDARD_CI_REPOSITORY_TYPE` (like `STANDARD_CI_SCM`) I chose it for pragmatic reasons - the term "repository type" is already used by NuGet in nuspec metadata. We don't strictly need to be consistent but for the same of less complexity it's better to be.