Single-File: Extraction configurations as build-time properties.
This commit changes the way the location where files in a bundle
are extracted can be influenced at build time.

Originally, they were written as properties in `runtimeconfig.json` file,
which is complex to handle at runtime, because the apphost needs to
extract resources before `hostfxr` process `runtimeconfig.json`
(because hostfxr itself needs to be extracted).

Instead the configurations are now build-time properties, which can be
written to and processed from the bundle-headers.