Single-File: Extraction configurations as build-time properties. (#76)
* Single-File: Extraction configurations update

This commit changes the way ExtractAllFiles property is configured. 

Originally, it was written as s property in `runtimeconfig.json` file,
which is complex to handle at runtime, because the apphost needs to
extract resources before `hostfxr` process `runtimeconfig.json`
(because hostfxr itself needs to be extracted).

Instead the configuration is now a build-time property, which can be
written to and processed from the bundle-headers.