Revert "Convert file format from JSON to YAML"
This reverts commit f55c61ee3e10c0fab79d4add9a62a6ec8cea7c95.

I decided to revert this back to JSON for the simple reason of consistency with the other file formats being used throughout .NET engineering (e.g. releases.json, runtime.json for RIDs, etc).  The main motivation for YAML was to have native support for comments.  JSON can still use comments but is technically unsupported in the standard.  This limits the general accessibility of the file but I would say that the vast majority of consumers would use the .NET library for interacting with the model due to the logic needed to interpret it.