Change the configuration to use file per arch
Change the way multi-arch configuration files are written. Use one file per architecture (instead of one file with multiple lines in it). This makes it much simpler for installers to read and write these configuration files. It also somewhat simplified the product as it doesn't need as complex parsing logic.

Minor cleanup in other areas of the doc.