Adjust compatibility rules for Apple TFMs.
Context: https://github.com/xamarin/xamarin-macios/issues/13087

We've decided to ship breaking changes with .NET 6 for iOS, macOS and tvOS,
and as such we won't be compatible with existing TFMs. All binaries will have
be recompiled for .NET 6.

I've modified compatibility scenarios that involve iOS or Mac Catalyst to use
Android instead (because the compatibility rules for Android won't change).